requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
deepmerge==2.0
distro==1.9.0
Django==5.1.6
django-bootstrap5==24.3
django-debug-toolbar==5.0.1
django-jquery==3.1.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jwt==1.3.1
numpy==1.26.4
openai==1.65.1
paho-mqtt==2.1.0
pillow==11.1.0
protobuf==4.25.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
qrcode==8.0
requests==2.32.3
sniffio==1.3.1
soundfile==0.12.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.3.0
sentry-sdk==2.19.2
python-decouple==3.8
gunicorn==23.0.0
python-dotenv==1.0.1
